+++ amotolani/cisco_fmc/plugins/modules/network.py
@@ -44,6 +44,12 @@ options:
       - IP address or FQDN of Cisco FMC.
     type: str
     required: true
+  fmc_timeout:
+    description:
+      - time to wait (seconds) for the API response from FMC.
+    type: int
+    default: 5
+    required: false
   value:
     description:
       - FMC network object value.
@@ -146,6 +152,7 @@ def main():
             network_type=dict(type='str', choices=['Host', 'Range', 'Network', 'FQDN'], required=True),
             value=dict(type='str', required=True),
             fmc=dict(type='str', required=True),
+            fmc_timeout=dict(type='int', required=False, default=5),
             username=dict(type='str', required=True),
             password=dict(type='str', required=True,  no_log=True),
             auto_deploy=dict(type='bool', default=False)
@@ -162,6 +169,7 @@ def main():
     network_type = module.params['network_type']
     value = module.params['value']
     fmc = module.params['fmc']
+    fmc_timeout = module.params['fmc_timeout']
     username = module.params['username']
     password = module.params['password']
     auto_deploy = module.params['auto_deploy']
@@ -248,7 +256,7 @@ def main():
         result = dict(failed=True, msg='Connection to FMC failed. Reason: {}'.format(err))
         module.exit_json(**result)
 
-    with FMC(host=fmc, username=username, password=password, autodeploy=auto_deploy) as fmc1:
+    with FMC(host=fmc, username=username, password=password, timeout=fmc_timeout, autodeploy=auto_deploy) as fmc1:
 
         # Instantiate Objects with values if valid ip, range or network address is provided
         if network_type == 'Host':